requirements.txt
# Web Server & API
fastapi
uvicorn[standard]
python-socketio
python-multipart

# AI & LLM integration
openai
langchain
langchain-openai
langchain-community
langchain-chroma

# Vector Database (Memory)
chromadb

# Audio Processing (The Ears)
openai-whisper
numpy

# Environment Variables
python-dotenv